unmuted narrowing conversion warning, converted all types from integer to float for ui classes to prevent hundreds of unecessary casts

+++ src/core/ui/button.hpp
@@ -61,7 +61,7 @@ namespace rl::ui {
 
     public:
         virtual void draw(vg::NVGcontext* ctx) override;
-        virtual ds::dims<i32> preferred_size(vg::NVGcontext* ctx) const override;
+        virtual ds::dims<f32> preferred_size(vg::NVGcontext* ctx) const override;
 
     private:
         bool handle_mouse_button_event(const ds::point<i32>& pt, Mouse::Button::ID button,
